Extract Rebuild library (#51709)
