+++ Compilers.slnf
@@ -30,6 +30,7 @@
       "src\\Compilers\\Core\\MSBuildTaskTests\\Microsoft.Build.Tasks.CodeAnalysis.UnitTests.csproj",
       "src\\Compilers\\Core\\MSBuildTask\\Microsoft.Build.Tasks.CodeAnalysis.csproj",
       "src\\Compilers\\Core\\Portable\\Microsoft.CodeAnalysis.csproj",
+      "src\\Compilers\\Core\\Rebuild\\Rebuild.csproj",
       "src\\Compilers\\Extension\\Roslyn.Compilers.Extension.csproj",
       "src\\Compilers\\Server\\VBCSCompilerTests\\VBCSCompiler.UnitTests.csproj",
       "src\\Compilers\\Server\\VBCSCompiler\\VBCSCompiler.csproj",
