+++ src/Compilers/Core/Rebuild/BlobReaderExtensions.cs
