+++ src/Compilers/Core/Rebuild/CompilationOptionsReader.cs
@@ -18,34 +18,10 @@ using Microsoft.Cci;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 
 namespace BuildValidator
 {
-    internal readonly struct SourceFileInfo
-    {
-        internal string SourceFilePath { get; }
-        internal SourceHashAlgorithm HashAlgorithm { get; }
-        internal byte[] Hash { get; }
-        internal SourceText? EmbeddedText { get; }
-        internal byte[]? EmbeddedCompressedHash { get; }
-
-        internal SourceFileInfo(
-            string sourceFilePath,
-            SourceHashAlgorithm hashAlgorithm,
-            byte[] hash,
-            SourceText? embeddedText,
-            byte[]? embeddedCompressedHash)
-        {
-            SourceFilePath = sourceFilePath;
-            HashAlgorithm = hashAlgorithm;
-            Hash = hash;
-            EmbeddedText = embeddedText;
-            EmbeddedCompressedHash = embeddedCompressedHash;
-        }
-    }
-
-    internal class CompilationOptionsReader
+    public class CompilationOptionsReader
     {
         // GUIDs specified in https://github.com/dotnet/runtime/blob/main/docs/design/specs/PortablePdb-Metadata.md#document-table-0x30
         public static readonly Guid HashAlgorithmSha1 = unchecked(new Guid((int)0xff1816ec, (short)0xaa5e, 0x4d10, 0x87, 0xf7, 0x6f, 0x49, 0x63, 0x83, 0x34, 0x60));
@@ -92,7 +68,7 @@ namespace BuildValidator
             return reader;
         }
 
-        public bool TryGetMetadataCompilationOptions([NotNullWhen(true)] out MetadataCompilationOptions? options)
+        internal bool TryGetMetadataCompilationOptions([NotNullWhen(true)] out MetadataCompilationOptions? options)
         {
             if (_metadataCompilationOptions is null && TryGetMetadataCompilationOptionsBlobReader(out var optionsBlob))
             {
@@ -103,7 +79,7 @@ namespace BuildValidator
             return options != null;
         }
 
-        public MetadataCompilationOptions GetMetadataCompilationOptions()
+        internal MetadataCompilationOptions GetMetadataCompilationOptions()
         {
             if (_metadataCompilationOptions is null)
             {
@@ -128,27 +104,6 @@ namespace BuildValidator
                 : Encoding.GetEncoding(encodingString);
 
             return encoding;
-        }
-
-        public ImmutableArray<SourceLink> GetSourceLinksOpt()
-        {
-            var sourceLinkUTF8 = GetSourceLinkUTF8();
-            if (sourceLinkUTF8 is null)
-            {
-                return default;
-            }
-
-            var parseResult = JsonConvert.DeserializeAnonymousType(Encoding.UTF8.GetString(sourceLinkUTF8), new { documents = (Dictionary<string, string>?)null });
-            return parseResult.documents.Select(makeSourceLink).ToImmutableArray();
-
-            static SourceLink makeSourceLink(KeyValuePair<string, string> entry)
-            {
-                // TODO: determine if this subsitution is correct
-                var (key, value) = (entry.Key, entry.Value); // TODO: use Deconstruct in .NET Core
-                var prefix = key.Remove(key.LastIndexOf("*"));
-                var replace = value.Remove(value.LastIndexOf("*"));
-                return new SourceLink(prefix, replace);
-            }
         }
 
         public byte[]? GetSourceLinkUTF8()
