+++ src/Compilers/Core/Rebuild/MetadataCompilationOptions.cs
