+++ src/Compilers/Core/Rebuild/SourceLink.cs
@@ -9,7 +9,7 @@ using System.Text;
 namespace BuildValidator
 {
     /// <summary>An entry in the source-link.json dictionary.</summary>
-    internal readonly struct SourceLink
+    public readonly struct SourceLink
     {
         public string Prefix { get; }
         public string Replace { get; }
