+++ src/Tools/BuildValidator/BuildValidator.csproj
@@ -14,6 +14,7 @@
     <ProjectReference Include="..\..\Compilers\Core\Portable\Microsoft.CodeAnalysis.csproj" />
     <ProjectReference Include="..\..\Compilers\CSharp\Portable\Microsoft.CodeAnalysis.CSharp.csproj" />
     <ProjectReference Include="..\..\Compilers\VisualBasic\Portable\Microsoft.CodeAnalysis.VisualBasic.vbproj" />
+    <ProjectReference Include="..\..\Compilers\Core\Rebuild\Rebuild.csproj" />
   </ItemGroup>
 
   <ItemGroup>
