+++ src/Tools/BuildValidator/BuildValidator.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.0.30905.15
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BuildValidator", "BuildValidator.csproj", "{F3EDD192-7E74-441D-B096-16C0FE2731E3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rebuild", "..\..\Compilers\Core\Rebuild\Rebuild.csproj", "{04CCBB7F-B1E3-4E76-A4A7-C648EBE77CBB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{F3EDD192-7E74-441D-B096-16C0FE2731E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F3EDD192-7E74-441D-B096-16C0FE2731E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F3EDD192-7E74-441D-B096-16C0FE2731E3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04CCBB7F-B1E3-4E76-A4A7-C648EBE77CBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04CCBB7F-B1E3-4E76-A4A7-C648EBE77CBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04CCBB7F-B1E3-4E76-A4A7-C648EBE77CBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04CCBB7F-B1E3-4E76-A4A7-C648EBE77CBB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
