+++ src/Tools/BuildValidator/LocalSourceResolver.cs
@@ -13,14 +13,6 @@ using Microsoft.Extensions.Logging;
 
 namespace BuildValidator
 {
-    internal record ResolvedSource(
-        string? OnDiskPath,
-        SourceText SourceText,
-        SourceFileInfo SourceFileInfo)
-    {
-        public string DisplayPath => OnDiskPath ?? ("[embedded]" + SourceFileInfo.SourceFilePath);
-    }
-
     internal class LocalSourceResolver
     {
         private readonly Options _options;
@@ -68,6 +60,22 @@ namespace BuildValidator
             }
 
             throw new FileNotFoundException(pdbDocumentPath);
+        }
+
+        internal ImmutableArray<ResolvedSource> ResolveSources(
+            ImmutableArray<SourceFileInfo> sourceFileInfos,
+            ImmutableArray<SourceLink> sourceLinks,
+            Encoding encoding)
+        {
+            _logger.LogInformation("Locating source files");
+
+            var sources = ImmutableArray.CreateBuilder<ResolvedSource>();
+            foreach (var sourceFileInfo in sourceFileInfos)
+            {
+                sources.Add(ResolveSource(sourceFileInfo, sourceLinks, encoding));
+            }
+
+            return sources.ToImmutable();
         }
     }
 }
